requirements.txt
Appium_Python_Client==2.11.1
Appium_Python_Client==2.11.1
pandas==2.0.3
selenium==4.11.2
